use lowercase class names for dict/list to mimic built-ins
these names make the api nicer since now usage of these classes looks
more like that of the built-in dict and list types.

however, this means that these names now shadow the "dict" and "list"
built-ins inside the module's code. deal with the consequences by
using "builtins.dict" and "builtins.list" when referring to the
built-in types, and by using internal aliases for the custom types to
avoid confusion.